refactor: step 1 - extract LoginCredentials record for spring UserDetails properties
