+++ dhis-2/dhis-services/dhis-service-core/src/main/java/org/hisp/dhis/security/DefaultLdapUserDetailsService.java
@@ -73,6 +73,6 @@ public class DefaultLdapUserDetailsService implements UserDetailsService {
     String password = "EXTERNAL_LDAP_" + CodeGenerator.generateCode(10);
     user.setPassword(password);
 
-    return userService.createUserDetails(user);
+    return userService.createUserDetails(user).getLoginCredentials();
   }
 }
